app: move the participant client app to evolution-frontend
** BREAKING CHANGE **

Surveys' client entry file now need to import the client app setup
function from evolution-frontend instead of legacy. For example:

```
import runClientApp from 'evolution-frontend/lib/apps/participant';
```
