+++ packages/evolution-frontend/package.json
@@ -55,6 +55,7 @@
     "popmotion": "^11.0.5",
     "react": "^19.0.0",
     "react-datepicker": "^7.6.0",
+    "react-dom": "^19.0.0",
     "react-i18next": "^12.3.1",
     "react-input-range": "^1.3.0",
     "react-loadable": "^5.5.0",
